function-calling-sql/duckdb-rag/streamlit/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
duckdb==0.9.2
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.41
groq==0.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==7.0.1
install==1.3.5
ipykernel==6.16.1
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.3
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
langchain==0.1.5
langchain-community==0.0.19
langchain-core==0.1.21
langsmith==0.0.87
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.20.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.2
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.2.1
nltk==3.8.1
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.23.4
openai==1.12.0
packaging==23.2
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pinecone-client==3.0.2
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==4.25.2
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
Send2Trash==1.8.0
sentence-transformers==2.3.1
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.25
sqlparse==0.4.4
stack-data==0.5.1
streamlit==1.31.0
sympy==1.12
tenacity==8.2.3
terminado==0.16.0
threadpoolctl==3.2.0
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.15.1
toml==0.10.2
toolz==0.12.1
torch==2.2.0
tornado==6.2
tqdm==4.66.1
traitlets==5.5.0
transformers==4.37.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
widgetsnbextension==4.0.3
yarl==1.9.4
zipp==3.17.0
